Function for checking intersection with image done
I implemented an addition to the check obstacles function. Now, the user
can pass a flag into the function which denotes whether or not it should
check for circle obstacles, or obstacles that are lines. The image is
turned into a series of lines using cv2.findContours, and the checkCollision
function checks whether or not the line from qHere to qThere intersects
with any of the contours that make up the obstacle. First the function
checks whether or not the line from qHere to qThere is a vertical line,
or if the obstacle contour is a vertical line. Then it checks whether
the slope between the line from qHere to qThere equals the slope of the
contour, and then it tries to find whether intersection between the two
lines falls inside one of the line segments.