messing with the way I make the start and goal
Right now in the mode with circular obstacles, I make the start and goal
positions first, and then when making each obstacle I check to see whether
or not the obstacle overlaps with either the start or the goal. When in
the mode with the NU "N", I do the exact opposite and make the start and goal
positions based off of where the obstacle is. I can't change the way I do
it in the mode with the N, so I have to change it int he mode with the
circles.